Add support to route registration and example use case
This architecture is based on having the routes and use cases organized
according to the app entities to be executed. In the case of an
application with users, we sould have a user folder inside /http and
another inside /app. This will allow us to have organized the boundaries
of our application.