Add connection to database
I have added the connection to a mysql database at application
bootstrap. To create this connection, i have created an interface called
Connection that contains a single method called connect.

To connect to the database, i have used gorm framework, which will allow
us to connect and make queries to database with a fast development
speed.

This connect method should return an instance of the gorm database. To
change the database engine, it would be enough for the moment to create
a struct that implements the Connection interface with the desired
engine.

To load the database configuratin, I have opted for the classic .env
file. Making use of the package godotenv.